fix(ops): argMax eager result must be a portable float tensor (native/wasm)
The eager DefaultCpuOps.argMax returned an Int32 payload inside a <FP32,Float>
tensor. JVM tolerated the boxing, but Kotlin/Native (linuxX64) and Wasm threw
`ClassCastException: Int cannot be cast to Float` when reading .data — a real
consumer bug, not just a test issue (surfaced by `allTests` in CI).

Fix: the eager kernel now materializes the indices as index-valued floats in the
input dtype, so the result is a consistent Tensor<T,V> readable on every target.
The traced/compiled form is unchanged — still a real i32 tensor (VoidTensorOps +
ArgMaxOperationsConverter), so the board vmfb path is unaffected.

Verified: DefaultCpuOpsArgMaxTest green on linuxX64 + jvm.

+++ skainet-backends/skainet-backend-cpu/src/commonTest/kotlin/sk/ainet/sk/ainet/exec/tensor/ops/DefaultCpuOpsArgMaxTest.kt
@@ -8,14 +8,13 @@ import sk.ainet.lang.tensor.Shape
 import sk.ainet.lang.tensor.Tensor
 import sk.ainet.lang.tensor.dsl.tensor
 import sk.ainet.lang.types.FP32
-import sk.ainet.lang.types.Int32
 
 class DefaultCpuOpsArgMaxTest {
     private val ctx = DirectCpuExecutionContext()
     private val ops get() = ctx.ops
 
-    // argMax returns an Int32-backed tensor (typed <FP32,Float> by the op signature);
-    // read the raw index via the underlying TensorData.
+    // Eager argMax returns index-valued floats in the input dtype (portable across JVM/native/wasm);
+    // read them back as ints.
     private fun idx(t: Tensor<FP32, Float>, vararg i: Int): Int = (t.data.get(*i) as Number).toInt()
 
     @Test
@@ -26,7 +25,7 @@ class DefaultCpuOpsArgMaxTest {
             val t = tensor<FP32, Float> { shape(2, 3) { init { v[it[0] * 3 + it[1]] } } }
             val r = ops.argMax(t, dim = -1)
             assertEquals(Shape(2), r.shape)
-            assertEquals<Any>(Int32::class, r.dtype) // runtime dtype is Int32 despite the <FP32,Float> static type
+            assertEquals<Any>(FP32::class, r.dtype) // eager result is the input dtype (index-valued floats)
             assertEquals(2, idx(r, 0))
             assertEquals(0, idx(r, 1))
         }
